requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy<2.0
timm
gradio
fastapi
uvicorn
pillow
google-genai
pydantic
python-multipart
requests

